nextjs_app/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv>=1.0.1
llama-cloud-services
Pillow==10.1.0 